requirements.txt
streamlit==1.54.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pandas==2.2.3
numpy==2.1.3
plotly==5.18.0
fpdf2==2.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
